検索画面の画面遷移を少なくした．スタイルはまだ指定していない．
